mark next address of ret

+++ linux-user/ptc.h
@@ -288,7 +288,7 @@ typedef struct {
   uint64_t *CallNext;
   uint32_t *isIndirectJmp;
   uint32_t *isDirectJmp;
-  uint32_t *isRet;
+  uint64_t *isRet;
   uint64_t *ElfStartStack;
 
 } PTCInterface;
